+++ src/main/java/org/opensearch/knn/common/KNNConstants.java
@@ -84,6 +84,7 @@ public class KNNConstants {
     public static final String TOP_LEVEL_ENGINE_FEATURE = "top_level_engine_feature";
 
     public static final String RADIAL_SEARCH_KEY = "radial_search";
+    public static final String NULL_K = "null_k";
     public static final String MODEL_VERSION = "model_version";
     public static final String QUANTIZATION_STATE_FILE_SUFFIX = "osknnqstate";
     public static final double ADC_CORRECTION_FACTOR = 2.0;
@@ -190,4 +191,12 @@ public class KNNConstants {
     public static final Integer INDEX_THREAD_QUANTITY_DEFAULT_LARGE = 4;
     public static final Integer INDEX_THREAD_QUANTITY_DEFAULT_SMALL = 1;
 
+    // mmr
+    public static final String MMR = "mmr";
+    public static final String DIVERSITY = "diversity";
+    public static final String CANDIDATES = "candidates";
+    public static final String VECTOR_FIELD_PATH = "vector_field_path";
+    public static final String VECTOR_FIELD_DATA_TYPE = "vector_field_data_type";
+    public static final String VECTOR_FIELD_SPACE_TYPE = "vector_field_space_type";
+    public static final String MMR_RERANK_CONTEXT = "mmr.rerank_context";
 }